pyproject.toml
[tool.poetry]
name = "company-management"
version = "0.1.0"
description = ""
authors = ["Viacheslav Kharchenko <vchkhr@gmail.com>"]
readme = "README.md"
packages = [{include = "company_management"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "4.0"
psycopg2 = "^2.9.6"
djangorestframework = "3.13"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
